refactor(backend): extract test crud module
Continues the incremental split of the crud monolith. The Test CRUD
functions move out of crud/__init__.py into a new crud/test.py; callers
switch to direct submodule imports. Pure move -- no logic changes.
