+++ tests/backend/crud/test_test_crud.py
@@ -20,6 +20,7 @@ from sqlalchemy.orm import Session
 
 from rhesis.backend.app import crud, models
 from rhesis.backend.app.constants import EXPLORER_REQUIREMENT_NAME
+from rhesis.backend.app.crud import test as test_crud
 from rhesis.backend.app.models.test import test_test_set_association
 from rhesis.backend.app.services import test_set as test_set_service
 from rhesis.backend.app.utils.crud_utils import count_items
@@ -39,7 +40,7 @@ class TestTestOperations:
         test_id = db_test.id
 
         # Test deletion - no mocking needed, test real requirement
-        result = crud.delete_test(
+        result = test_crud.delete_test(
             db=test_db, test_id=test_id, organization_id=test_org_id, user_id=authenticated_user_id
         )
 
@@ -57,7 +58,7 @@ class TestTestOperations:
         """Test deletion of non-existent test"""
         fake_test_id = uuid.uuid4()
 
-        result = crud.delete_test(
+        result = test_crud.delete_test(
             db=test_db,
             test_id=fake_test_id,
             organization_id=test_org_id,
@@ -122,7 +123,7 @@ class TestTestOperations:
         assert seeded_test_set is not None
         assert seeded_test_set.attributes["metadata"]["requirements"] == ["Compliance"]
 
-        result = crud.update_test(
+        result = test_crud.update_test(
             db=test_db,
             test_id=db_test.id,
             test={"requirement_id": robustness.id},
@@ -149,7 +150,7 @@ class TestTestOperations:
         with patch(
             "rhesis.backend.app.services.test_set.update_test_set_attributes"
         ) as mock_refresh:
-            result = crud.update_test(
+            result = test_crud.update_test(
                 db=test_db,
                 test_id=db_test_minimal.id,
                 test={"priority": 3},
@@ -202,7 +203,7 @@ class TestTestOperations:
         )
         test_db.flush()
 
-        crud.update_test(
+        test_crud.update_test(
             db=test_db,
             test_id=db_test.id,
             test={"requirement_id": robustness.id},
@@ -225,7 +226,7 @@ class TestTestOperations:
 @pytest.mark.unit
 @pytest.mark.crud
 class TestBulkDeleteTests:
-    """🧪 crud.bulk_delete_tests"""
+    """🧪 test_crud.bulk_delete_tests"""
 
     def _make_test(self, test_db, test_org_id, authenticated_user_id, requirement_id, test_set_id):
         db_test = models.Test(
@@ -268,7 +269,9 @@ class TestBulkDeleteTests:
 
         # 2 tests in set A, 1 test in set B
         tests_a = [
-            self._make_test(test_db, test_org_id, authenticated_user_id, requirement.id, test_set_a.id)
+            self._make_test(
+                test_db, test_org_id, authenticated_user_id, requirement.id, test_set_a.id
+            )
             for _ in range(2)
         ]
         test_b = self._make_test(
@@ -279,7 +282,7 @@ class TestBulkDeleteTests:
         with patch(
             "rhesis.backend.app.services.test_set.update_test_set_attributes"
         ) as mock_refresh:
-            result = crud.bulk_delete_tests(
+            result = test_crud.bulk_delete_tests(
                 db=test_db,
                 test_ids=all_ids,
                 organization_id=test_org_id,
@@ -306,7 +309,7 @@ class TestBulkDeleteTests:
         """A nonexistent id in the batch is reported back, not silently dropped."""
         fake_id = uuid.uuid4()
 
-        result = crud.bulk_delete_tests(
+        result = test_crud.bulk_delete_tests(
             db=test_db,
             test_ids=[db_test_minimal.id, fake_id],
             organization_id=test_org_id,
@@ -319,7 +322,7 @@ class TestBulkDeleteTests:
     def test_empty_ids_is_a_noop(
         self, test_db: Session, test_org_id: str, authenticated_user_id: str
     ):
-        result = crud.bulk_delete_tests(
+        result = test_crud.bulk_delete_tests(
             db=test_db, test_ids=[], organization_id=test_org_id, user_id=authenticated_user_id
         )
         assert result == {"deleted_ids": [], "not_found_ids": []}
@@ -378,7 +381,7 @@ class TestBulkDeleteTests:
         with patch(
             "rhesis.backend.app.services.test_set.update_test_set_attributes"
         ) as mock_refresh:
-            result = crud.bulk_delete_tests(
+            result = test_crud.bulk_delete_tests(
                 db=test_db,
                 test_ids=[own_test.id, foreign_test.id],
                 organization_id=test_org_id,
@@ -404,7 +407,7 @@ class TestBulkDeleteTests:
 @pytest.mark.unit
 @pytest.mark.crud
 class TestGetTestsExcludesExplorer:
-    """crud.get_tests must omit explorer tests (general test list API)."""
+    """test_crud.get_tests must omit explorer tests (general test list API)."""
 
     def test_get_tests_excludes_explorer_rows(
         self, test_db: Session, test_org_id: str, authenticated_user_id: str
@@ -421,7 +424,7 @@ class TestGetTestsExcludesExplorer:
         test_db.add_all([regular, explorer])
         test_db.commit()
 
-        results = crud.get_tests(
+        results = test_crud.get_tests(
             test_db,
             organization_id=str(test_org_id),
             user_id=str(authenticated_user_id),
