+++ tests/backend/jobs/test_batch_contract_resolution.py
@@ -45,7 +45,9 @@ class TestResolveContractLazy:
 
         with (
             patch("rhesis.backend.app.database.get_db_with_tenant_variables") as mock_get_db,
-            patch("rhesis.backend.app.crud.get_test", return_value=fresh_test) as mock_get_test,
+            patch(
+                "rhesis.backend.app.crud.test.get_test", return_value=fresh_test
+            ) as mock_get_test,
             patch(
                 _RESOLVE_MULTI_TURN_CONTRACT,
                 return_value=({"prohibited_behavior": ["X"]}, True),
@@ -71,7 +73,7 @@ class TestResolveContractLazy:
 
         with (
             patch("rhesis.backend.app.database.get_db_with_tenant_variables") as mock_get_db,
-            patch("rhesis.backend.app.crud.get_test", return_value=None),
+            patch("rhesis.backend.app.crud.test.get_test", return_value=None),
         ):
             mock_get_db.return_value.__enter__.return_value = MagicMock()
             mock_get_db.return_value.__exit__.return_value = False
@@ -129,7 +131,7 @@ class TestRunMultiTurnContractThreading:
             patch(
                 _RESOLVE_MULTI_TURN_CONTRACT, return_value=({"prohibited_behavior": ["X"]}, True)
             ),
-            patch("rhesis.backend.app.crud.get_test", return_value=test),
+            patch("rhesis.backend.app.crud.test.get_test", return_value=test),
             patch("rhesis.backend.app.database.get_db_with_tenant_variables") as mock_get_db,
             patch("rhesis.backend.jobs.execution.penelope_target.BackendEndpointTarget"),
         ):
@@ -159,7 +161,7 @@ class TestRunMultiTurnContractThreading:
 
         with (
             patch(_RESOLVE_MULTI_TURN_CONTRACT, return_value=(None, False)),
-            patch("rhesis.backend.app.crud.get_test", return_value=test),
+            patch("rhesis.backend.app.crud.test.get_test", return_value=test),
             patch("rhesis.backend.app.database.get_db_with_tenant_variables") as mock_get_db,
             patch("rhesis.backend.jobs.execution.penelope_target.BackendEndpointTarget"),
         ):
