+++ tests/backend/security/test_organization_filtering.py
@@ -10,11 +10,12 @@ import uuid
 import pytest
 from sqlalchemy.orm import Session
 
-from rhesis.backend.app import crud, models
+from rhesis.backend.app import models
 from rhesis.backend.app.crud import endpoint as endpoint_crud
 from rhesis.backend.app.crud import model as model_crud
 from rhesis.backend.app.crud import prompt as prompt_crud
 from rhesis.backend.app.crud import task as task_crud
+from rhesis.backend.app.crud import test as test_crud
 from rhesis.backend.app.crud import test_result as test_result_crud
 from rhesis.backend.app.crud.metric import create_metric, get_metric
 from rhesis.backend.app.crud.test_run import get_test_run
@@ -111,7 +112,7 @@ class TestCrudOrganizationFiltering:
         test_db.commit()
 
         # User from org1 should be able to access the test
-        result_org1 = crud.get_test(
+        result_org1 = test_crud.get_test(
             test_db, test_obj.id, organization_id=str(org1.id), user_id=str(user1.id)
         )
         assert result_org1 is not None
@@ -119,7 +120,7 @@ class TestCrudOrganizationFiltering:
         assert str(result_org1.organization_id) == str(org1.id)
 
         # User from org2 should NOT be able to access the test
-        result_org2 = crud.get_test(
+        result_org2 = test_crud.get_test(
             test_db, test_obj.id, organization_id=str(org2.id), user_id=str(user2.id)
         )
         assert result_org2 is None
@@ -611,7 +612,7 @@ class TestCrudParameterValidation:
         # getattr/hasattr lookup would skip them and check nothing.
         crud_functions = [
             ("get_task", task_crud.get_task),
-            ("get_test", crud.get_test),
+            ("get_test", test_crud.get_test),
             ("get_test_result", test_result_crud.get_test_result),
             ("get_test_run", get_test_run),
             ("get_endpoint", endpoint_crud.get_endpoint),
